+++ src/controllers/BulkEditController.php
@@ -5,7 +5,7 @@
  * Bulk edit entries
  *
  * @link      https://venveo.com
- * @copyright Copyright (c) 2018 Venveo
+ * @copyright Copyright (c) 2018-2019 Venveo
  */
 
 namespace venveo\bulkedit\controllers;
@@ -15,9 +15,7 @@ use craft\records\Element;
 use craft\records\Field;
 use craft\web\Controller;
 use craft\web\Response;
-use venveo\bulkedit\assetbundles\bulkeditscreen\BulkEditScreenAsset;
-use venveo\bulkedit\BulkEdit;
-use venveo\bulkedit\BulkEdit as Plugin;
+use venveo\bulkedit\Plugin;
 use venveo\bulkedit\queue\jobs\SaveBulkEditJob;
 use venveo\bulkedit\records\EditContext;
 use venveo\bulkedit\records\History;
@@ -25,8 +23,6 @@ use venveo\bulkedit\services\BulkEdit as BulkEditService;
 use yii\web\BadRequestHttpException;
 
 /**
- * https://craftcms.com/docs/plugins/controllers
- *
  * @author    Venveo
  * @package   BulkEdit
  * @since     1.0.0
@@ -38,25 +34,29 @@ class BulkEditController extends Controller
      *
      * @return Response
      * @throws BadRequestHttpException if not a valid request
-     * @throws \Twig_Error_Loader
-     * @throws \yii\base\Exception
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
+     * @throws \craft\errors\SiteNotFoundException
      */
     public function actionGetFields(): Response
     {
-        $this->requireLogin();
         $this->requirePostRequest();
         $this->requireAcceptsJson();
 
         $site = Craft::$app->getSites()->getCurrentSite();
         $elementIds = Craft::$app->getRequest()->getRequiredParam('elementIds');
+        $elementType = Craft::$app->getRequest()->getRequiredParam('elementType');
         $requestId = Craft::$app->getRequest()->getRequiredParam('requestId');
 
         /** @var BulkEditService $service */
         $service = Plugin::$plugin->bulkEdit;
-        $layouts = $service->getFieldLayoutsForElementIds($elementIds);
+        $fields = $service->getFieldsForElementIds($elementIds, $elementType);
+
+
         $view = \Craft::$app->getView();
         $modalHtml = $view->renderTemplate('venveo-bulk-edit/elementactions/BulkEdit/_fields', [
-            'layouts' => $layouts,
+            'fieldWrappers' => $fields,
             'bulkedit' => $service,
             'elementIds' => $elementIds,
             'site' => $site
@@ -72,46 +72,59 @@ class BulkEditController extends Controller
         $responseData['headHtml'] = $view->getHeadHtml();
         $responseData['footHtml'] = $view->getBodyHtml();
 
-
         return $this->asJson($responseData);
     }
 
-    public function actionGetEditScreen(): Response
+    /**
+     * @return \yii\web\Response
+     * @throws BadRequestHttpException
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
+     */
+    public function actionGetEditScreen()
     {
-        $this->requireLogin();
         $this->requirePostRequest();
         $this->requireAcceptsJson();
 
-        $elementIds = array_values(Craft::$app->getRequest()->getRequiredParam('elementIds'));
+
+        $elementIds = Craft::$app->getRequest()->getRequiredParam('elementIds');
         $requestId = Craft::$app->getRequest()->getRequiredParam('requestId');
         $siteId = Craft::$app->getRequest()->getRequiredParam('siteId');
-        $fieldIds = array_values(array_filter(Craft::$app->getRequest()->getRequiredParam('fieldIds')));
+        $fields = Craft::$app->getRequest()->getRequiredParam('fields');
+
+        // Pull out the enabled fields
+        $enabledFields = [];
+        foreach ($fields as $fieldId => $field) {
+            if ($field['enabled']) {
+                $enabledFields[$fieldId] = $field;
+            }
+        }
+
 
         $site = Craft::$app->getSites()->getSiteById($siteId);
         if (!$site) {
             throw new \Exception('Site does not exist');
         }
 
-        $fields = Field::findAll($fieldIds);
-        if (count($fields) !== count($fieldIds)) {
+
+        $fields = Field::findAll(array_keys($enabledFields));
+        if (count($fields) !== count($enabledFields)) {
             throw new \Exception('Could not find all fields requested');
         }
 
+        $elementIds = explode(',', $elementIds);
         $elements = Element::findAll($elementIds);
         if (count($elements) !== count($elementIds)) {
             throw new \Exception('Could not find all elements requested');
         }
-
-
-        $view = Craft::$app->getView();
-        $view->registerAssetBundle(BulkEditScreenAsset::class);
 
         try {
             $fieldModels = [];
             /** @var Field $field */
             foreach ($fields as $field) {
                 $fieldModel = \Craft::$app->fields->getFieldById($field->id);
-                if ($fieldModel && BulkEdit::$plugin->bulkEdit->isFieldSupported($fieldModel)) {
+                if ($fieldModel && Plugin::$plugin->bulkEdit->isFieldSupported($fieldModel)) {
                     $fieldModels[] = $fieldModel;
                 }
             }
@@ -119,13 +132,20 @@ class BulkEditController extends Controller
             throw $e;
         }
 
-        $baseEntry = null;
         $view = \Craft::$app->getView();
+
+        // We've gotta register any asset bundles - this won't actually be rendered
+        foreach ($fieldModels as $fieldModel) {
+            $view->renderPageTemplate('_includes/field', [
+                'field' => $fieldModel,
+                'required' => false
+            ]);
+        }
 
         $modalHtml = $view->renderTemplate('venveo-bulk-edit/elementactions/BulkEdit/_edit', [
             'fields' => $fieldModels,
             'elementIds' => $elementIds,
-            'baseElement' => $baseEntry,
+            'fieldData' => $enabledFields,
             'site' => $site
         ]);
         $responseData = [
@@ -140,16 +160,26 @@ class BulkEditController extends Controller
         return $this->asJson($responseData);
     }
 
-    public function actionSaveContext(): Response
+    /**
+     * @return \yii\web\Response
+     * @throws BadRequestHttpException
+     * @throws \yii\db\Exception
+     */
+    public function actionSaveContext()
     {
-        $this->requireLogin();
         $this->requirePostRequest();
         $this->requireAcceptsJson();
 
         $elementIds = Craft::$app->getRequest()->getRequiredParam('elementIds');
         $siteId = Craft::$app->getRequest()->getRequiredParam('siteId');
-        $fieldIds = array_values(Craft::$app->getRequest()->getRequiredParam('fieldIds'));
+//        $fieldIds = array_values(Craft::$app->getRequest()->getRequiredParam('fieldIds'));
+        $fieldMeta = array_values(Craft::$app->getRequest()->getRequiredParam('fieldMeta'));
 
+        $fieldStrategies = [];
+        foreach ($fieldMeta as $field) {
+            $fieldStrategies[$field['id']] = $field['strategy'];
+        }
+        $fieldIds = array_keys($fieldStrategies);
         $fields = Field::findAll($fieldIds);
 
         $values = Craft::$app->getRequest()->getBodyParam('fields', []);
@@ -161,11 +191,13 @@ class BulkEditController extends Controller
                     $fieldId = $field->id;
                 }
             }
-            if ($fieldId === null) {
+            if (!isset($fieldId)) {
                 throw new \Exception('Failed to locate field');
             }
             $keyedFieldValues[$fieldId] = $value;
         }
+
+        $elementIds = explode(',', $elementIds);
 
         $context = new EditContext();
         $context->ownerId = \Craft::$app->getUser()->getIdentity()->id;
@@ -177,6 +209,8 @@ class BulkEditController extends Controller
         $rows = [];
         foreach ($elementIds as $elementId) {
             foreach ($fieldIds as $fieldId) {
+                $strategy = $fieldStrategies[$fieldId] ?? 'replace';
+
                 $rows[] = [
                     'pending',
                     $context->id,
@@ -185,11 +219,12 @@ class BulkEditController extends Controller
                     (int)$siteId,
                     '[]',
                     \GuzzleHttp\json_encode($keyedFieldValues[$fieldId]),
+                    $strategy
                 ];
             }
         }
 
-        $cols = ['status', 'contextId', 'elementId', 'fieldId', 'siteId', 'originalValue', 'newValue'];
+        $cols = ['status', 'contextId', 'elementId', 'fieldId', 'siteId', 'originalValue', 'newValue', 'strategy'];
         \Craft::$app->db->createCommand()->batchInsert(History::tableName(), $cols, $rows)->execute();