+++ src/services/BulkEdit.php
@@ -13,8 +13,9 @@ namespace venveo\bulkedit\services;
 use Craft;
 use craft\base\Component;
 use craft\base\Element;
+use craft\base\Field;
 use craft\base\FieldInterface;
-use craft\elements\Entry;
+use craft\events\RegisterComponentTypesEvent;
 use craft\fields\BaseRelationField;
 use craft\fields\Checkboxes;
 use craft\fields\Color;
@@ -28,6 +29,14 @@ use craft\fields\RadioButtons;
 use craft\fields\Table;
 use craft\fields\Url;
 use craft\records\FieldLayout;
+use craft\redactor\Field as RedactorField;
+use venveo\bulkedit\base\AbstractElementTypeProcessor;
+use venveo\bulkedit\elements\processors\AssetProcessor;
+use venveo\bulkedit\elements\processors\CategoryProcessor;
+use venveo\bulkedit\elements\processors\EntryProcessor;
+use venveo\bulkedit\elements\processors\ProductProcessor;
+use venveo\bulkedit\elements\processors\UserProcessor;
+use venveo\bulkedit\models\FieldWrapper;
 use venveo\bulkedit\records\EditContext;
 use venveo\bulkedit\records\History;
 
@@ -38,38 +47,63 @@ use venveo\bulkedit\records\History;
  */
 class BulkEdit extends Component
 {
+    public const STRATEGY_REPLACE = 'replace';
+    public const STRATEGY_MERGE = 'merge';
+    public const STRATEGY_SUBTRACT = 'subtract';
+
+    public const EVENT_REGISTER_ELEMENT_PROCESSORS = 'registerElementProcessors';
+    public const EVENT_REGISTER_SUPPORTED_FIELDS = 'registerSupportedFields';
+
     /**
      * Get all distinct field layouts from a set of elements
      *
      * @param $elementIds
+     * @return FieldWrapper[] fields
+     * @throws \Exception
      */
-    public function getFieldLayoutsForElementIds($elementIds)
+    public function getFieldsForElementIds($elementIds, $elementType)
     {
-        $layouts = FieldLayout::find()
-            ->select('fieldlayouts.*')
-            ->distinct(true)
-            ->limit(null)
-            ->from('{{%fieldlayouts}} fieldlayouts')
-            ->leftJoin('{{%elements}} elements', '[[elements.fieldLayoutId]] = [[fieldlayouts.id]]')
-            ->where(['IN', '[[elements.id]]', $elementIds])
-            ->all();
-
-        $layoutsModels = [];
-        /** @var FieldLayout $layout */
-        foreach($layouts as $layout) {
-            $layoutsModels[$layout->id] = ['fields' => \Craft::$app->fields->getFieldsByLayoutId($layout->id)];
+        // Works for entries
+        $processor = $this->getElementTypeProcessor($elementType);
+        if (!$processor) {
+            throw new \Exception('Unable to process element type');
         }
-        return $layoutsModels;
+        $layouts = $processor::getLayoutsFromElementIds($elementIds);
+
+        $fields = [];
+        /** @var FieldLayout $layout */
+        foreach ($layouts as $layout) {
+            $layoutFields = \Craft::$app->fields->getFieldsByLayoutId($layout->id);
+            /** @var Field $layoutField */
+            foreach ($layoutFields as $layoutField) {
+                if (!array_key_exists($layoutField->handle, $fields)) {
+                    $fieldWrapper = new FieldWrapper();
+                    $fieldWrapper->field = $layoutField;
+                    $fieldWrapper->layouts[] = $layout;
+                    $fields[$layoutField->handle] = $fieldWrapper;
+                } else {
+                    $fields[$layoutField->handle]->layouts[] = $layout;
+                }
+            }
+        }
+        return $fields;
     }
 
-    public function getBulkEditContextFromId($id)
+
+    /**
+     * @param $id
+     * @return EditContext|null
+     */
+    public function getBulkEditContextFromId($id): ?EditContext
     {
         return EditContext::findOne($id);
     }
 
     /**
+     * Gets all unique elements from incomplete bulk edit tasks
+     *
      * @param EditContext $context
-     * @return array
+     * @return int[] pending element IDs
      */
     public function getPendingElementIdsFromContext(EditContext $context): array
     {
@@ -82,6 +116,8 @@ class BulkEdit extends Component
     }
 
     /**
+     * Gets all pending bulk edit tasks
+     *
      * @param EditContext $context
      * @return \yii\db\ActiveQueryInterface
      */
@@ -91,6 +127,8 @@ class BulkEdit extends Component
     }
 
     /**
+     * Gets all pending tasks for a particular element
+     *
      * @param EditContext $context
      * @param $elementId
      * @return \yii\db\ActiveQueryInterface
@@ -106,33 +144,54 @@ class BulkEdit extends Component
      * Takes an array of history changes for a particular element and saves it to that element.
      *
      * @param $historyItems
-     * @param \craft\base\Element $element
-     * @return \craft\base\Element
+     * @param Element $element
+     * @return Element
      * @throws \Throwable
      * @throws \yii\base\Exception
      */
-    public function processHistoryItemsForElement($historyItems, \craft\base\Element $element)
+    public function processHistoryItemsForElement($historyItems, Element $element): ?Element
     {
+        // We'll process the entire element in a transaction to help avoid problems
         $transaction = Craft::$app->getDb()->beginTransaction();
         try {
+            /** @var History $historyItem */
             foreach ($historyItems as $historyItem) {
-                $historyItem->originalValue = \GuzzleHttp\json_encode($element->getFieldValue($historyItem->field->handle));
+                $fieldHandle = $historyItem->field->handle;
+                $newValue = \GuzzleHttp\json_decode($historyItem->newValue, true);
+                $originalValue = $element->getFieldValue($historyItem->field->handle);
+                $historyItem->originalValue = \GuzzleHttp\json_encode($originalValue);
                 $historyItem->status = 'completed';
-                $element->setFieldValue($historyItem->field->handle, \GuzzleHttp\json_decode($historyItem->newValue));
+                $field = \Craft::$app->fields->getFieldByHandle($fieldHandle);
+                switch ($historyItem->strategy) {
+                    case self::STRATEGY_REPLACE:
+                        $element->setFieldValue($fieldHandle, $newValue);
+                        break;
+                    case self::STRATEGY_MERGE:
+                        if ($field && $field instanceof BaseRelationField) {
+                            $ids = $originalValue->ids();
+                            $ids = array_merge($ids, $newValue);
+                            $element->setFieldValue($fieldHandle, $ids);
+
+                        } else {
+                            throw new \Exception("Can't merge field: " . $fieldHandle);
+                        }
+                        break;
+                    case self::STRATEGY_SUBTRACT:
+                        if ($field && $field instanceof BaseRelationField) {
+                            $ids = $originalValue->ids();
+                            $ids = array_diff($ids, $newValue);
+                            $element->setFieldValue($fieldHandle, $ids);
+                        } else {
+                            throw new \Exception("Can't subtract field: " . $fieldHandle);
+                        }
+                        break;
+
+                }
                 $historyItem->save();
                 Craft::info('Saved history item', __METHOD__);
             }
             $element->setScenario(Element::SCENARIO_ESSENTIALS);
             \Craft::$app->elements->saveElement($element, false);
-
-            switch (get_class($element)) {
-                case Entry::class:
-                    // Save a revision
-                    \Craft::$app->entryRevisions->saveVersion($element);
-                    break;
-                default:
-                    break;
-            }
 
             Craft::info('Saved element', __METHOD__);
             $transaction->commit();
@@ -161,11 +220,78 @@ class BulkEdit extends Component
             MultiSelect::class
         ];
 
+        // Add support for redactor
+        if (\Craft::$app->getPlugins()->isPluginEnabled('redactor')) {
+            $supportedFields[] = RedactorField::class;
+        }
+
+        $event = new RegisterComponentTypesEvent();
+        $event->types = &$supportedFields;
+        $this->trigger(self::EVENT_REGISTER_SUPPORTED_FIELDS, $event);
+
         foreach ($supportedFields as $fieldItem) {
             if ($field instanceof $fieldItem) {
                 return true;
             }
         }
         return false;
+    }
+
+    /**
+     * Gets an array of values for supported strategies on field types
+     * @param FieldInterface $field
+     * @return array
+     */
+    public function getSupportedStrategiesForField(FieldInterface $field)
+    {
+        $availableStrategies = [
+            ['value' => self::STRATEGY_REPLACE, 'label' => 'Replace']
+        ];
+
+        if ($field instanceof BaseRelationField) {
+            $availableStrategies[] = ['value' => self::STRATEGY_MERGE, 'label' => 'Merge'];
+            $availableStrategies[] = ['value' => self::STRATEGY_SUBTRACT, 'label' => 'Subtract'];
+        }
+
+        return $availableStrategies;
+    }
+
+    /**
+     * Retrieves the processor for a type of element. The processor determines how to do things like get the field
+     * layout.
+     * @param $elementType
+     * @return AbstractElementTypeProcessor
+     * @throws \ReflectionException
+     */
+    public function getElementTypeProcessor($elementType)
+    {
+        $processors = [
+            EntryProcessor::class,
+            UserProcessor::class,
+            CategoryProcessor::class,
+            AssetProcessor::class,
+        ];
+
+        if (Craft::$app->plugins->isPluginInstalled('commerce')) {
+            $processors[] = ProductProcessor::class;
+        }
+
+        $event = new RegisterComponentTypesEvent();
+        $event->types = &$processors;
+        $this->trigger(self::EVENT_REGISTER_ELEMENT_PROCESSORS, $event);
+
+        $processorsKeyedByClass = [];
+        foreach ($processors as $processor) {
+            $reflection = new \ReflectionClass($processor);
+            /** @var AbstractElementTypeProcessor $instance */
+            $instance = $reflection->newInstanceWithoutConstructor();
+            $type = $instance::getType();
+            $processorsKeyedByClass[$type] = $processor;
+        }
+
+        if (array_key_exists($elementType, $processorsKeyedByClass)) {
+            return $processorsKeyedByClass[$elementType];
+        }
+        return null;
     }
 }